requirements/dev.txt
-r base.txt
black==24.1.1
django-debug-toolbar==4.3.0